Add calibration error display to GUI
Enhancements to the camera control GUI were made to display the calibration error. Previously, the GUI didn't have a specific section to display the calibration error after the calibration process. These changes allow users to directly see the calibration error on the GUI, improving the user experience. Also, the precise value of the calibration error is displayed, allowing for more easily interpreted results.