TestEnable camera calibration and snap frame buttons
The code that was previously commented out has been enabled. This change allows the camera to perform a calibration and snap a single frame. This modification enables more functionalities for the camera, such as the ability to conduct a calibration check and the ability to snap a frame on demand. Also, in the 'SetFrameRate' function, the 'StopLive()' and 'StartLive()' calls have been commented out to avoid unwanted camera stoppage and restart during frame rate modification.