Add termination option to toggle_calibration_capture()
A termination parameter was added to the toggle_calibration_capture() of the camera_control_GUI_improved.py. Previously, the method would only switch off only if the calibration_capture_toggle_status was True. However, this change allows the method to also switch off if termination is set to True. This is useful for ensuring that all recording threads complete their process before the calibration is toggled off, avoiding potential concurrent processing issues. The check for recording thread completion now also excludes the current thread, reinforcing the safety of thread management.