Improve handling of camera calibration and frame processing
This commit optimizes the way frames are processed during camera calibration in the GUI. It includes a synchronization barrier to wait for all the calibration tasks to complete before proceeding, thus, avoiding eventual inconsistencies. This also makes sure the calibration capture toggle status is correctly reset after each calibration iteration.

Additionally, this introduces a more efficient handling of frame processing. It now ensures all remaining frames in the queue are processed for each thread, and only cleared when they meet the frame processing threshold. This addresses a potential issue where some frames could and object detection data. As an output, the complete set of detection rows are dumped into a pickle file.