Refactor draw_axis method to improve camera rotation and translation
The draw_axis method in camera_control_GUI_improved.py has been updated to take rotation and translation of the camera into account. Previously, it only used the camera matrix and distortion coefficients. Now, rotation and translation parameters have been added to provide more accurate image rendering. This change will impact how images are shown in the GUI to the user, leading to a more realistic rendering of the camera's perspective. This could potentially enhance the user's ability to control the camera and interpret the camera's viewpoint.