Add parameter to draw_reprojection_on_thread method
The "draw_reprojection_on_thread" method was updated to take an additional parameter "num", allowing us to specify a unique value for each thread. This change was needed to accommodate the logic that different threads have different identifiers, which significantly improves the execution and performance of threaded tasks. The respective call in "camera_control_GUI_improved.py" was also updated to pass