Add video_output attribute to Camera class
A new attribute, 'vid_out', has been added to the Camera class in the ic_camera.py file. This attribute is utilized to store the video output from the camera, which was previously not stored. The addition of this attribute improves the effectiveness of the camera functionality and allows for potentially enhanced video processing capabilities.