"Refactor frame writing process in Camera Controller"
This commits refactor the camera frame acquisition and writing process in `ic_camera.py`. Changes are made to improve concurrency and to prevent blocking due to I/O delay.

A `write` method is replaced with two methods: `acquire_frame` and `write_frame`. `acquire_frame` does not write the frame directly to the video output anymore, instead, it appends the frame, time data, and frame number to a deque buffer. `write_frame` takes top frame from the deque buffer and writes it into the video output. These changes are introduced to increase the performance and introduce thread safety while writing frames into the video output.

Another modification is added a mechanism to keep track of the recording status and a separate thread for processing frames to ensure video writing doesn't block frame capturing.