"Added setup name to video naming and GUI
The update introduces the use of experimental setup name in creating the base names for the videos. This helps in better video identification and organization especially when multiple setups are used. Additionally, a field for the setup name was added to the GUI under the Video info setting, enhancing user experience. The setup name is initialized to 'Test' by default."