Add functionality to display coordinates on mouse click in camera image
This commit implements the feature to display coordinates of a clicked point in the image captured by the camera. Now, when the user clicks on a point in the image window, the x and y coordinates of the clicked point are shown on the console as well as on the image window. This was achieved by calling a callback function `click_event` on mouse click events in the OpenCV window showing the camera image. The callback receives the position of the mouse click and uses the cv2.putText method to display the coordinates on the image. Imported libraries cv2 and PIL to handle the image and its display.