Add tracking coordinates to video trigger setup
This commit includes updates to camera control functions, specifically `set_up_video_trigger()`, `release_video_file()` and `set_params()`. A new argument, trackingCoords, has been added to their lists of parameters. This allows users to setup the desired object tracking coordinates in the video. The trackingCoords argument holds values for x and y coordinates of the object to be tracked. If no tracking point is defined, the camera will proceed as usual, but if a tracking point is provided, it will also capture and store the pixel value of the tracking point. The changes have been incorporated in the GUI trigger setup as well.