Fixed byte string parameters in camera ROI settings
This commit changes the arguments passed to the SetPropertyAbsoluteValue function from string to byte string. The former method was causing compatibility issues as the library used (likely for interacting with the hardware camera) was expecting byte string arguments instead. This led to unwanted and incorrect behavior whenever the ROI (Region of Interest) of the camera was set. By converting the arguments to byte string, this problem is resolved and the function operates as expected.