Uncommented certain lines and commented out `set_ROI()` in ic_camera.py, and commented out the camera flipping section in camera_control_GUI_improved.py
In the ic_camera.py file, certain lines that were initially commented have been uncommented to ensure proper initialization and operation of the camera. Action has been shifted into creating and starting the camera in the necessary format with the appropriate filters, instead of setting an ROI section. Moreover, added a line to flip the frames vertically.

In the camera_control_GUI_improved.py, the block of code responsible for flipping the cameras vertically has been commented out. This change is made since these operations are now handled individually within the ic_camera.py file. As a result, the camera flipping procedure is more centralized, reducing redundancy and chances for discrepancy.