Updated add_filters to allow custom values
The function add_filters() in the file ic_camera.py has been updated to accept custom values for top, left, height, and width. Previously, the function was limited to only using values of self.crop for these parameters. This change provides greater flexibility when calling this function.

The reopening and starting of the cam in flip_vertical() function were also added for better handling of video format and filters.