"Refactor camera format setup and add delay after format change"
This commit includes two substantial changes to the camera settings functionality. Firstly, instead of using a simple string format for setting camera parameters, the 'set_formats' method now accepts individual 'width' and 'height' parameters. This allows more flexibility and control over the camera settings from the GUI.

Secondly, a delay has been added after setting the camera format. This was necessary because a common problem was that the camera did not have enough time to adjust to the new format before the format was used, leading to issues in capturing images. This delay ensures the camera has time to properly setup the format.

On top of these, a redundant 'set_video_ROI' method has been removed from the ic_camera class.