Change default overwrite status and number of cams data type
Changed the default value of the overwrite parameter in the setup_trigger_recording method from True to False to prevent accidental data loss. Also, the data type for the number_of_cams variable was adjusted from string to integer to align with the expected data types elsewhere in the code and to improve input validation.