Set static values for offset configurations in GUI
The offset configurations for both x and y axes on the GUI used to be set dynamically depending on the frame dimensions. This was causing problems when the frame dimensions were not standard. Now, the offset values have been hardcoded to 1440 for x and 1080 for y to provide uniformity and resolve issues with non-standard frame sizes.