Revised calibration controls and updated row loading output
Commented out the recalibrate and update calibration button states, as these are not necessary once calibration has been completed. Added an additional print message to show the number of rows loaded from detections.pickle file. This gives clearer feedback to the user about the progress and state of the calibration process.