Add timeout feature to video recording
Added a new feature to handle video recording timeouts in the camera control module. A timeout status and start time were added to track whether a recording should be terminated after a set duration. This prevents excessively long recordings.

In the GUI code, a cycle count variable was used to append a cycle number to each new file generated after a timeout occurs, ensuring no overwrites.

Added a new function, save_trigger_recording_on_thread, to handle the saving of videos when a timeout occurs. This function also prepares the next video file for recording.

The timeout feature is used to save the videos automatically in between triggering sessions