Improve error handling in camera settings
Added exception handling in the 'set_exposure' and 'set_gain' functions in the ic_camera module. In previous versions, if setting the exposure or gain failed, this was silently ignored which made debugging difficult. Now, when an error occurs, the traceback and exception type and value are printed to console, which will make it easier to diagnose and fix issues.