Disable camera timeout functionality
The timeout functionality for camera control is disabled. This includes stopping the camera recording when no frames have been captured for a certain period of time and deleting the videos when no frame was captured since the last trigger train. This will allow continuous recording without getting interrupted by the timeout.