Adjust f-string formatting and increase frame buffer size
This commit updates the 'ic_camera.py' script. Firstly, an f-string format is added to a print statement for visual consistency and to ensure correct variable interpolation. Secondly, the size of the frame buffer deque is increased from 50 to 250, providing a larger buffer for frame storage which helps in handling more frames in memory while processing videos, thereby improving video processing speed.